Finished the 6th activity
This is the challenge I expected and I was challenged
The instruction explained too much which made it more confusing for me
but read it again then I was engaged. The button part was easy but
the paragraph made me look twice but its the same as the button.
The last part was the real deal as I intended to only move the
selectors but having the child class was enough hint to solve the
acitivty