Finished 2nd activity
I commit first my answer then later check the solution so I can have
comparison what how I do on these activities.
This rack up my brain as I think I forgot most of the lesson.
The only part I got confused was the card and its contents gap as I was
trying with the padding and margin it is either to big or none so I just
stick with the margin of the contents in which I think do the job.