Reviewed answer in the 2nd activity
So Apparently I did'nt forgot the lessons but I did not use them
properly haha. The button part was the only one I got correctly
I got confused on the margin and padding relationship of the
card padding and margin of the contents which I ignored somewhat
and only focused on the margin. I forgot also that you can use
part of the margin top bottom etc.. which was significally the
answer to this activity