Finished 2nd activity after seeing solution
I was dumbfounded. The big reason on why I used marign and padding at
the ul element was I cannot center the items inside which the only
solution I think of was adding a selector of both div parents and style
it align-items. I cannot still comprehend on why the link items was
styled by the align items in the header selector. Is it that it
recognized as a child or something happended which I was not aware of.