Finish activity 3 before solution
It is the same with the 2nd activity but it challenge by making you
create the container in which the items will based on.

If anyone is reading this If I didn't commit the answer after the
I saw the solution it means that I did similarly with the solution or
my answer was the same with the solution.