Finish 3rd activity with solution
This one made me realize again that anything can be a solution to a
certain design I am confident that I attained the similar design but my
takeway with the solution was the code will be easier to read and able
put more elements easier.

If anyone is reading this here is my explantion of my solution
I immediately identified the child Logo, UL, and right-items in which I
put margin-left auto as a  way to attain the desired design. Then put
the align-items in the header as per understading with the 2nd activity
and added gap and padding for the spacing.