finish 4th activity before solution
This is fun. Getting all the planned way in youyr mind play out in the
broswer is rewarding. I was supposed to use the body as the container
then it occur that  was a rookie mistake so I made a container for the
title and plants then I nested container on the plants img and text
which make them 2 child for the container class then I put container on
each plant with text. Making use of the flex-direction was easy to
put where you want the items to be. For some reason I think I over coded
and think this would be done more efficiently