Finish 4th activity after solution
My guess was right it could be done more efficiently even though the
design was achieved they are too many code and the solution made it
flawless. My first mistake was not using css and making all the elements
into flex. My takeaway in this was the harmony of flex and block which I
thought they cannot be used together. The other that bothers me was
using the gap for the spacing which I don't mind but isn't like the use
of space-around etc in will be absolete or I'm just overthinking.