Finish 5th activity before solution
My method was contain the header,close,text, and buttons then nest the
header and close and as well the buttons choices. Apply the flex-shrink
to 0 that is important. I'm not sure on the spacing part even though
they almost look exactly the same but they don't. I think the spacing
was incorrectly applied to my solution.