Finish 5th activity after solution
I overdone the nesting part and overthink the properties that some
elements has. I was shocked at the close button to be put on the header
then I learned that the text inside can be items which I didn't take
advantage of. THe align-items in the header was a sly one as you cannot
see much on the preview but it is a good technique. I was over fixated
using gap in which everytime I see a space gap was my goto and forget
about margin and padding. Add also the display block - The buttons at
end I put them in a container as a safeway to secure them but the
container did not need a display parameter to attain the design.