Finish 6th activity before solution
This is fun. Without the last hint I may not be able to finish this in
a short time. I was skeptical first to use it as I thought the solution
was use margin auto on both header & footer thus I was proven wrong. As
I was going through the aligning the items made me realized the flex-
direction was better and using block will be dead end. Styling was the
easy part and the only part that bothers me was the links. I am
contemplating since I put the header & footer selector in css but stick
to ul as it still did the job.