Finish 7th activity before solution
This is the hardest one. I can't seem to wrap around my head the flex-
wrap but with a little touch up on spacing I think I got the desire
design but I base on the spacing I think I used the wrong properties
for heigh and width of cards but nevertheless It is similar. Also note
that emoji doest work so I don't know if the side links was supposed to
be aligned perfectly but that is my answer. The header gave me also a
headache but I think my solution was valid based on what have been done
previously.