Finish 7th activity with solution
First of all english was my third language and partly my fault as I skim
on readme. My first solution on header was correct I was just wrong on
the padding in which I thought all over but only it need on the screen.
Th Footer was correct. The links was a goofy one casuse I thought the
literal size was the solution but it refers to the font-size lol but the
solution was correct. From the size I thought of separting the emoji and
text. The wrap is easy but the spacing is what I don't get but only
minor revision from the original. What I did wrong on wrap was to add
height which I thought was the solution , I even put max on both but
I still got similar result onyl spacing was noticable.